requirements.txt
Django>=4.2.0,<5.0
python-decouple>=3.6
pywebpush>=1.14.0
mysqlclient>=2.1.0
requests>=2.28.0
reportlab>=4.0.0
Pillow>=10.0.0
qrcode>=7.4.0
openpyxl>=3.1.0
pythainlp>=5.0.0
django-summernote>=0.8.20
beautifulsoup4>=4.12.0